requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.6
exceptiongroup==1.1.3
googlesearch-python==1.2.3
h11==0.14.0
idna==3.4
outcome==1.3.0
pycparser==2.21
PySocks==1.7.1
requests==2.31.0
selenium==4.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.22.2
trio-websocket==0.11.1
urllib3==2.0.7
wsproto==1.2.0
